requirements.txt
certifi==2025.11.12
charset-normalizer==3.4.4
dominate==2.9.1
idna==3.11
imageio==2.35.1
joblib==1.4.2
lazy_loader==0.4
networkx==3.1
numpy==1.24.4
opencv-python==4.12.0.88
packaging==25.0
pillow==10.4.0
PyWavelets==1.4.1
requests==2.32.4
scikit-image==0.21.0
scikit-learn==1.3.2
scipy==1.10.1
threadpoolctl==3.5.0
tifffile==2023.7.10
torch==1.13.0+cu117
torchaudio==0.13.0+cu117
torchvision==0.14.0+cu117
tqdm==4.67.1
typing_extensions==4.13.2
urllib3==2.2.3
